chore: Move InputActionWrapper to ux-editor-v3
